==> 2-Import and Clean BO Order Data.Rmd <==
---
title: "Import & Clean BO Order Data"
output: html_notebook
---

```{r include = FALSE}
#import data
library(readxl)
bo <- read_excel("~/Library/Mobile Documents/com~apple~CloudDocs/Documents/EM Research/Data Analysis/RVPv3.4/RawData/rvpcxr.orders.Oct17toSept19.xlsx", 
    col_types = c("text", "date", "numeric", 
        "date", "date", "date", "text", "text", 
        "text", "text", "text"))

track <- readRDS(file = "track.clean.Rda")

#load packages
library(tidyverse)
library(lubridate)
library(psych)

#export raw data
saveRDS(bo, file = "bo.raw.Rda")
```

##Distributions
###Dispo
```{r}
bo$Dispo_BO <- factor(bo$Dispo_BO) #convert to factors
bo %>% count(Dispo_BO)
```

###Order Status
```{r}
bo$OrderStatus <- factor(bo$OrderStatus)
bo %>% count(OrderStatus)
```

##Mnemonic Name
```{r}
bo$MnemonicName <- factor(bo$MnemonicName)
bo %>% count(MnemonicName)
```

```{r}
#export cleaned data
saveRDS(bo, file = "dfs/bo.clean.Rda")

#split off CXR df
cxr <- bo %>% subset((grepl("XR",bo$MnemonicName, ignore.case=FALSE))) %>% droplevels()
saveRDS(cxr, file = "dfs/cxr.Rda")

#split off RVP df
rvp <- bo %>% subset((!grepl("XR",bo$MnemonicName, ignore.case=FALSE))) %>% droplevels()
saveRDS(rvp, file = "dfs/rvp.Rda")
```
