==> 3-Match ED Tracking with RVP.Rmd <==
---
title: "Match ED Tracking w/ RVP"
output: html_notebook
---

```{r}
#Import & libraries
cxr <- readRDS(file = "dfs/cxr.Rda")
rvp <- readRDS(file = "dfs/rvp.Rda")
track <- readRDS(file = "dfs/track.dedup.Rda")

library(tidyverse)
library(lubridate)
library(psych)
```


```{r}
# create df with only encounters in which RVP ordered and tracking info, joined via FIN
track$FIN %>% n_distinct() #checks if any duplicates

rvp.track <- merge(x = track,
                   y = rvp,
                   by = "FIN",
                   all = F)

```


```{r}
#list column names
names(rvp.track)
```

```{r}
#rearrange, drop, and rename columns

#col_order <- c("FIN", "Sex", "DOB", "Age", "RFV", "RFV_ED", "Admit_DateTime", "Arrival_DT", "Reg_DT", "Checkin_DT", "TriageStart_DT", "MDAssigned_DT", "PinOrder_DT", "TransferOrder_DT", "BedAssign_DT", "DischargeDecision_DT", "Checkout_DT", "Discharge_DateTime", )

rvp.track.clean <- rvp.track %>% select(FIN, Sex, DOB, Age,
                     RFV, RFV_ED,
                     Admit_DateTime, Arrival_DT, Reg_DT, Checkin_DT, TriageStart_DT, MDAssigned_DT, 
                     PinOrder_DT, TransferOrder_DT, BedAssign_DT, DischargeDecision_DT, Checkout_DT, Discharge_DateTime,
                     OrderID, MnemonicName, OrderStatus, Order_DateTime, OrderComplete_DateTime, 
                     Dispo_ED, Dispo_BO,
                     DoorToTriage.mins, DoorToBed.mins, DoorToMDAssigned.mins, BedToMDAssigned.mins, DoorToDispo.mins, MDAssignedToDischargeDecision.mins, 
                     CheckinToCheckout.mins)

rvp.track.clean <-  rvp.track.clean %>% rename(RVP_OrderID = OrderID, 
                                               RVP_OrderStatus = OrderStatus, 
                                               RVP_OrderDT = Order_DateTime, 
                                               RVP_OrderCompleteDT = OrderComplete_DateTime)
                     

saveRDS(rvp.track.clean, file = "dfs/rvp.track.clean.Rda") #export
```
