==> 4- Match CXR Data with ED Tracking and RVP.Rmd <==
---
title: "Match CXR Data w/ Tracking and RVP"
output: html_notebook
---


```{r}
#Import & libraries
cxr <- readRDS(file = "dfs/cxr.Rda")
rvp.track <- readRDS(file = "dfs/rvp.track.clean.Rda")

library(tidyverse)
library(lubridate)
library(psych)
```


```{r}
names(cxr)
```


```{r}
# keep only necessary CXR columns and rename
cxr.clean <- cxr %>% select(FIN, 
               XR_OrderID = OrderID, 
               XR_Type = MnemonicName, 
               XR_OrderStatus = OrderStatus, 
               XR_OrderDT = Order_DateTime, 
               XR_OrderCompleteDT = OrderComplete_DateTime)

saveRDS(cxr.clean, file = "dfs/cxr.clean.Rda") #export

cxr <- cxr.clean
```

```{r}
levels(cxr$XR_Type)
cxr %>% count(XR_Type)
cxr %>% count(XR_OrderStatus)
cxr$FIN %>% n_distinct()

cxr %>% count(XR_Type)
```


```{r}
#change cxr Order Status levels
cxr$XR_OrderStatus <- cxr$XR_OrderStatus %>% factor(levels = c("Completed", "Discontinued", "Canceled", "Voided"))
levels(cxr$XR_OrderStatus)

#change cxr type levels so that 2 views is first (and in descending order of count)
cxr$XR_Type <- cxr$XR_Type %>% factor(levels = c("XR Chest 2 Views Frontal/Lat", "XR Chest 1 View Frontal", "XR Chest 1 View PA", "XR Chest 3 Views Frontal/Lat/Apical", "XR Chest Routine"))
levels(cxr$XR_Type)
                                            
```

```{r}
#Check CXR duplicates
cxr.dup <- cxr[duplicated(cxr$FIN)|duplicated(cxr$FIN, fromLast = TRUE),] %>% #show encounters with duplicate FIN
  arrange(FIN, XR_OrderStatus, XR_Type) #sort so Completed and 2 View shows up first (and doesn't get filtered)
```

```{r}
#apply above duplicate sort to full set
cxr <- cxr %>% arrange(FIN, XR_OrderStatus, XR_Type) #sort so Completed and 2 View shows up first (and doesn't get filtered)

#remove duplicates
cxr.dedup <- cxr %>% distinct(FIN, .keep_all = TRUE) #remove duplicates (removes 2nd+ row)

saveRDS(cxr.dedup, file = "dfs/cxr.dedup.Rda")
```

```{r}
#match all rvp.track with corresponding CXR; do NOT include CXR with FINs not in rvp.track

rvp.track.cxr <- merge(x = rvp.track,
                       y = cxr.dedup,
                       by = "FIN",
                       all.x = TRUE)
```


```{r}
rvp.track.cxr %>% count(MnemonicName)
```


```{r}
#create binary if CXR was completed, regardless of type
rvp.track.cxr$XR_Completed <- ifelse(rvp.track.cxr$XR_OrderStatus == "Completed", 1, 0) %>% factor()
rvp.track.cxr$XR_Completed <- fct_explicit_na(rvp.track.cxr$XR_Completed, na_level = "0")
rvp.track.cxr$XR_Completed <- rvp.track.cxr$XR_Completed %>% factor(levels = c(1,0),
                                                                    labels = c("XR Completed", "No XR Completed"))

rvp.track.cxr %>% count(XR_Completed)
```


```{r}
table(rvp.track.cxr$MnemonicName, rvp.track.cxr$XR_Completed)
```


```{r}
#export
saveRDS(rvp.track.cxr, file = "dfs/rvp.track.cxr.Rda")
```
