==> 5- Split Data Set by MnemonicName.Rmd <==
---
title: "Split Data Set by MnemonicName"
output: html_notebook
---


```{r}
#Import & libraries
rvp <- readRDS(file = "dfs/rvp.track.cxr.Rda")

library(tidyverse)
library(lubridate)
library(psych)
```


```{r}
rvp %>% count(MnemonicName)
```

```{r}
table(rvp$MnemonicName, rvp$RVP_OrderStatus)
```


```{r}
#show encounters with duplicate FIN & MnemonicName
rvp.dup <- rvp[duplicated(rvp[c("FIN", "MnemonicName")])|
                duplicated(rvp[c("FIN", "MnemonicName")], 
                fromLast =TRUE), ] %>% 
  arrange(FIN, MnemonicName, RVP_OrderStatus, RVP_OrderDT)

rvp.dup %>% count(MnemonicName)
```

```{r}
#remove encounters w/ duplicate EncounterIDs and MnemonicName (same order 2x)
rvp <- rvp %>% arrange(FIN, MnemonicName, RVP_OrderStatus, RVP_OrderDT) %>% distinct(FIN, MnemonicName, .keep_all = TRUE)

rvp %>% count(MnemonicName)
```

```{r}
#export dedup
saveRDS(rvp, file = "dfs/rvp.track.cxr.dedup.Rda")
```

Extract datasets of each test only (in case of duplicates), export as Rda
```{r}
#Flu POCT
only.poct.flu <- rvp %>% subset(
  (grepl("Influenza",rvp$MnemonicName, ignore.case=FALSE)))
saveRDS(only.poct.flu, file="dfs/only.poct.flu.Rda")

#RSV POCT
only.poct.rsv <- rvp %>% subset(
  (grepl("RSV",rvp$MnemonicName, ignore.case=FALSE)))
saveRDS(only.poct.rsv, file="dfs/only.poct.rsv.Rda")

#Strep A POCT
only.poct.strep <- rvp %>% subset(
  (grepl("Strep A \\(POCT\\)",rvp$MnemonicName, ignore.case=FALSE)))
saveRDS(only.poct.strep, file="dfs/only.poct.strep.Rda")

#RespPCR
only.lab.pcr <- rvp %>% subset(
  (grepl("Respiratory",rvp$MnemonicName, ignore.case=FALSE)))
saveRDS(only.lab.pcr, file="dfs/only.lab.pcr.Rda")

#Rapid Strep
only.lab.strep <- rvp %>% subset(
  (grepl("Rapid Strep A",rvp$MnemonicName, ignore.case=FALSE)))
saveRDS(only.lab.strep, file="dfs/only.lab.strep.Rda")
```
