==> 6- Merge POCT vs Lab Data Sets.Rmd <==
---
title: "Merge POCT vs Lab Data Sets"
output: html_notebook
---

```{r}
#Import & libraries
only.poct.flu <- readRDS(file = "dfs/only.poct.flu.Rda")
only.poct.rsv <- readRDS(file = "dfs/only.poct.rsv.Rda")
only.poct.strep <- readRDS(file = "dfs/only.poct.strep.Rda")
only.lab.pcr <- readRDS(file = "dfs/only.lab.pcr.Rda")
only.lab.strep <- readRDS(file = "dfs/only.lab.strep.Rda")

library(tidyverse)
library(lubridate)
library(psych)
```


Create dataframe of POCT Flu vs. Resp PCR, removing all encounters in which both tests were ordered
```{r}
#create df of the intersecting encounters (semijoin returns rows of POCT flu w/ matched FIN Lab PCR)
only.poct.flu %>% semi_join(only.lab.pcr, by = "FIN") -> flu.intersect

#create df of POCT flu w/o those encounters (via antijoin)
only.poct.flu %>% anti_join(flu.intersect, by = "FIN") -> flu.poct.dedup
saveRDS(flu.poct.dedup, file="dfs/flu.poct.dedup.Rda")

#create df of Resp PCR w/o those encounters (via antijoin)
only.lab.pcr %>% anti_join(flu.intersect, by = "FIN") -> flu.lab.dedup
saveRDS(flu.lab.dedup, file="dfs/flu.lab.dedup.Rda")

#merge the two dataframes
rbind(flu.lab.dedup, flu.poct.dedup) %>% arrange(Arrival_DT) %>% droplevels() -> df.flu

#export df
saveRDS(df.flu,file="dfs/df.flu.Rda")

df.flu %>% count(MnemonicName)
```

Create dataframe of POCT Flu vs. Resp PCR, before and after
```{r}
#subset lab pcr
lab.pcr.before <- only.lab.pcr %>% filter(Admit_DateTime < as.Date("2018-10-01"))

#export
saveRDS(lab.pcr.before, file = "dfs/lab.pcr.before.Rda")

#import POCT flu orders without concurrent Resp PCR orders
flu.poct.dedup <- readRDS(file="dfs/flu.poct.dedup.Rda")

#merge the two dataframes
rbind(lab.pcr.before, flu.poct.dedup) %>% arrange(Arrival_DT) %>% droplevels() -> df.flu.ba

#export df
saveRDS(df.flu.ba,file="dfs/df.flu.beforeafter.Rda")

```

Create dataframe of POCT Strep vs. Rapid Strep A Antigen, removing all encounters in which both tests were ordered
```{r}
#create df of the intersecting encounters (semijoin returns rows of POCT Strep w/ matched FIN Lab Strep)
only.poct.strep %>% semi_join(only.lab.strep, by = "FIN") -> strep.intersect

#create df of POCT flu w/o those encounters (via antijoin)
only.poct.strep %>% anti_join(strep.intersect, by = "FIN") -> strep.poct.dedup
saveRDS(strep.poct.dedup, file="dfs/strep.poct.dedup.Rda")

#create df of Resp PCR w/o those encounters (via antijoin)
only.lab.strep %>% anti_join(strep.intersect, by = "FIN") -> strep.lab.dedup
saveRDS(strep.lab.dedup, file="dfs/strep.lab.dedup.Rda")

strep.lab.dedup <- readRDS(file = "dfs/strep.lab.dedup.Rda")
strep.poct.dedup <- readRDS(file="dfs/strep.poct.dedup.Rda")

#merge the two dataframes
rbind(strep.lab.dedup, strep.poct.dedup) %>% arrange(Arrival_DT) %>% droplevels() -> df.strep

#export df
saveRDS(df.strep,file="dfs/df.strep.Rda")

df.strep %>% count(MnemonicName)
```
