==> 7 - Flu Analysis.Rmd <==
---
title: "7 - Flu Analysis"
output: html_notebook
---


```{r}
#Import & libraries
flu <- readRDS(file = "dfs/df.flu.Rda")

library(tidyverse)
library(lubridate)
library(psych)
```

##Determining Time and Duration Variables

```{r}
#number of NAs in DischargeDecision
is.na(flu$DischargeDecision_DT) %>% table(flu$MnemonicName)
```


```{r}
#number of NAs in BedAssign
is.na(flu$BedAssign_DT) %>% table(flu$MnemonicName)
```

Most of the time (all but 85 encounters), Discharge_DateTime is earlier than Checkout_DT
```{r}
#compare Checkout DT vs Discharge_DateTime (greater = later)
ifelse(flu$Checkout_DT < flu$Discharge_DateTime, 1, 0) %>% sum()
```

Most of the time (all but 659 encounters), MDAssigned_DT is after BedAssign_DT
```{r}
#compare BedAssign vs MDAssign (greater = later)
ifelse(flu$BedAssign_DT > flu$MDAssigned_DT, 1, 0) %>% sum(na.rm = TRUE) 
```


```{r}
#Compute Bed Assigned to Discharge Decision
flu$BedToDispo.mins<- interval(ymd_hms(flu$BedAssign_DT), ymd_hms(flu$DischargeDecision_DT)) %>% as.duration()/dminutes(1)

#number of NAs in BedToDispo
is.na(flu$BedToDispo.mins) %>% table(flu$MnemonicName)
```

```{r}
# Compute Bed Assigned to Discharge
flu$BedToDischarge.mins<- interval(ymd_hms(flu$BedAssign_DT), ymd_hms(flu$Discharge_DateTime)) %>% as.duration()/dminutes(1)

# number of NAs in BedToDischarge
is.na(flu$BedToDischarge.mins) %>% table(flu$MnemonicName)
```


```{r}
#Compute AdmitToDischarge
flu$AdmitToDischarge.mins<- interval(ymd_hms(flu$Admit_DateTime), ymd_hms(flu$Discharge_DateTime)) %>% as.duration()/dminutes(1)

#number of NAs in AdmitToDischarge
is.na(flu$AdmitToDischarge.mins) %>% table(flu$MnemonicName)
```

Admit To Discharge Stats
```{r}
do.call("rbind", describeBy(flu$AdmitToDischarge.mins, flu$MnemonicName))
```

Checkin To Checkout Stats
```{r}
do.call("rbind", describeBy(flu$CheckinToCheckout.mins, flu$MnemonicName))
```

```{r}
#export flu with added duration vars
saveRDS(flu, file = "dfs/df.flu.durs.Rda")
```


Sex Comparison
```{r}
flu$Sex %>% table(flu$MnemonicName)
chisq.test(flu$Sex, flu$MnemonicName, correct=FALSE)
```

##Remove Outliers

Remove AdmitToDischarge Outliers, check stats
```{r}
#Remove extreme outliers by group (removes values >3IQR+3Q, LOS <15 min); 
flu1 <-flu %>% group_by(MnemonicName) %>% filter(!(AdmitToDischarge.mins > (quantile(AdmitToDischarge.mins, .75, na.rm=T)+3*IQR(AdmitToDischarge.mins, na.rm=TRUE))))
flu1 <- flu1 %>% group_by(MnemonicName) %>% filter(!(AdmitToDischarge.mins < 15))

#Remove all outliers by group (removes values >1.5IQR+3Q, LOS <15 min); 
flu2 <-flu %>% group_by(MnemonicName) %>% filter(!(AdmitToDischarge.mins > (quantile(AdmitToDischarge.mins, .75, na.rm=T)+1.5*IQR(AdmitToDischarge.mins, na.rm=TRUE))))
flu2 <- flu2 %>% group_by(MnemonicName) %>% filter(!(AdmitToDischarge.mins < 15))

do.call("rbind", describeBy(flu1$AdmitToDischarge.mins, flu1$MnemonicName))
do.call("rbind", describeBy(flu2$AdmitToDischarge.mins, flu2$MnemonicName))

do.call("rbind", describeBy(flu1$Age, flu1$MnemonicName))
```

Remove CheckinToCheckout Outliers, check stats
```{r}
#Remove extreme outliers by group (removes values >3IQR+3Q, LOS <15 min); 
flu3 <-flu %>% group_by(MnemonicName) %>% filter(!(CheckinToCheckout.mins > (quantile(CheckinToCheckout.mins, .75, na.rm=T)+3*IQR(CheckinToCheckout.mins, na.rm=TRUE))))
flu3 <- flu3 %>% group_by(MnemonicName) %>% filter(!(CheckinToCheckout.mins < 15))

#Remove all outliers by group (removes values >1.5IQR+3Q, LOS <15 min); 
flu4 <-flu %>% group_by(MnemonicName) %>% filter(!(CheckinToCheckout.mins > (quantile(CheckinToCheckout.mins, .75, na.rm=T)+1.5*IQR(CheckinToCheckout.mins, na.rm=TRUE))))
flu4 <- flu4 %>% group_by(MnemonicName) %>% filter(!(CheckinToCheckout.mins < 15))

do.call("rbind", describeBy(flu3$CheckinToCheckout.mins, flu3$MnemonicName))
do.call("rbind", describeBy(flu4$CheckinToCheckout.mins, flu4$MnemonicName))
```

##Traditional Flu CXR Analysis

Comparison table
```{r}
table(flu$MnemonicName, flu$XR_Completed)
```

Chi-square
```{r}
chisq.test(flu$MnemonicName, flu$XR_Completed, correct = FALSE)
```


##Traditional Flu LOS Analysis

LOS Stats
```{r}
do.call("rbind", describeBy(flu1$AdmitToDischarge.mins, flu1$MnemonicName))
```

Shapiro Test - failed
```{r}
with(flu1, shapiro.test(AdmitToDischarge.mins[MnemonicName == "Influenza (POCT)"]))
with(flu1, shapiro.test(AdmitToDischarge.mins[MnemonicName == "Respiratory PCR Panel"]))
```

Wilcoxon test - significant
```{r}
wilcox.test(AdmitToDischarge.mins ~ MnemonicName, data = flu1, alternative = "two.sided", var.equal = FALSE, conf.int = TRUE, conf.level = 0.95)
```

Bootstrapped Median CI
```{r}
library(rcompanion)
groupwiseMedian(data=flu1, AdmitToDischarge.mins ~ MnemonicName, bca = FALSE, percentile = TRUE, R = 1000)
```

##Traditional Flu Bed to Dispo Analysis

Bed to Dispo Stats
```{r}
do.call("rbind", describeBy(flu$BedToDispo.mins, flu$MnemonicName))
```

Shapiro Test - failed
```{r}
with(flu, shapiro.test(BedToDispo.mins[MnemonicName == "Influenza (POCT)"]))
with(flu, shapiro.test(BedToDispo.mins[MnemonicName == "Respiratory PCR Panel"]))
```

Wilcoxon test - significant
```{r}
wilcox.test(BedToDispo.mins ~ MnemonicName, data = flu, alternative = "two.sided", var.equal = FALSE, conf.int = TRUE, conf.level = 0.95)
```


##Modified ILI Flu Setup

Subset by modified ILI
```{r}
#fever, cough, sore throat, flu, cold, body*ache, chills
flu.ili <- flu %>% subset(
  grepl("fever",flu$RFV, ignore.case=TRUE)| 
  grepl("cough",flu$RFV, ignore.case=TRUE)|
  grepl("sore.*throat",flu$RFV, ignore.case=TRUE)|
  grepl("throat.*pain",flu$RFV, ignore.case=TRUE)|
  grepl("flu",flu$RFV, ignore.case=TRUE)|
  grepl("cold",flu$RFV, ignore.case=TRUE)|
  grepl("body.*ache",flu$RFV, ignore.case=TRUE)|
  grepl("chills",flu$RFV, ignore.case=TRUE)
)

do.call("rbind", describeBy(flu.ili$AdmitToDischarge.mins, flu.ili$MnemonicName))
```

Remove AdmitToDischarge Outliers, check stats
```{r}
#Remove extreme outliers by group (removes values >3IQR+3Q, LOS <15 min); 
flu.ili1 <-flu.ili %>% group_by(MnemonicName) %>% filter(!(AdmitToDischarge.mins > (quantile(AdmitToDischarge.mins, .75, na.rm=T)+3*IQR(AdmitToDischarge.mins, na.rm=TRUE))))
flu.ili1 <- flu.ili1 %>% group_by(MnemonicName) %>% filter(!(AdmitToDischarge.mins < 15))

#Remove all outliers by group (removes values >1.5IQR+3Q, LOS <15 min); 
flu.ili2 <-flu.ili %>% group_by(MnemonicName) %>% filter(!(AdmitToDischarge.mins > (quantile(AdmitToDischarge.mins, .75, na.rm=T)+1.5*IQR(AdmitToDischarge.mins, na.rm=TRUE))))
flu.ili2 <- flu.ili2 %>% group_by(MnemonicName) %>% filter(!(AdmitToDischarge.mins < 15))

do.call("rbind", describeBy(flu.ili1$AdmitToDischarge.mins, flu.ili1$MnemonicName))
do.call("rbind", describeBy(flu.ili2$AdmitToDischarge.mins, flu.ili2$MnemonicName))

do.call("rbind", describeBy(flu.ili1$Age, flu.ili1$MnemonicName))
```


##Modified ILI CXR Analysis

Comparison table
```{r}
table(flu.ili$MnemonicName, flu.ili$XR_Completed)
```

Chi-square
```{r}
chisq.test(flu.ili$MnemonicName, flu.ili$XR_Completed, correct = FALSE)
```

```{r}
#export flu.ili
saveRDS(flu.ili, file = "dfs/df.flu.ili.Rda")
```

##Modified ILI LOS Analysis

LOS Stats
```{r}
do.call("rbind", describeBy(flu.ili1$AdmitToDischarge.mins, flu.ili1$MnemonicName))
```

LOS Shapiro Test - failed
```{r}
with(flu.ili1, shapiro.test(AdmitToDischarge.mins[MnemonicName == "Influenza (POCT)"]))
with(flu.ili1, shapiro.test(AdmitToDischarge.mins[MnemonicName == "Respiratory PCR Panel"]))
```

LOS Wilcoxon test - significant
```{r}
wilcox.test(AdmitToDischarge.mins ~ MnemonicName, data = flu.ili, alternative = "two.sided", var.equal = FALSE, conf.int = TRUE, conf.level = 0.95)
```

LOS Bootstrapped Median CI
```{r}
library(rcompanion)
groupwiseMedian(data=flu.ili, AdmitToDischarge.mins ~ MnemonicName, bca = FALSE, percentile = TRUE, R = 1000)
```


##Modified ILI BedToDispo Analysis

Bed to Dispo Stats
```{r}
do.call("rbind", describeBy(flu.ili$BedToDispo.mins, flu.ili$MnemonicName))
```

Shapiro Test - failed
```{r}
with(flu.ili, shapiro.test(BedToDispo.mins[MnemonicName == "Influenza (POCT)"]))
with(flu.ili, shapiro.test(BedToDispo.mins[MnemonicName == "Respiratory PCR Panel"]))
```

Wilcoxon test - significant
```{r}
wilcox.test(BedToDispo.mins ~ MnemonicName, data = flu.ili, alternative = "two.sided", var.equal = FALSE, conf.int = TRUE, conf.level = 0.95)
```
