==> 8 - Strep Analysis.Rmd <==
---
title: "Strep Analysis"
output: html_notebook
---

```{r}
#Import & libraries
strep <- readRDS(file = "dfs/df.strep.Rda") 

#reorder MnemonicName levels
strep$MnemonicName <- strep$MnemonicName %>% factor(levels = c("Strep A (POCT)","Rapid Strep A Antigen"))

library(tidyverse)
library(lubridate)
library(psych)
```

##Compute Time Duration Variables

```{r}
#Compute Bed Assigned to Discharge Decision
strep$BedToDispo.mins<- interval(ymd_hms(strep$BedAssign_DT), ymd_hms(strep$DischargeDecision_DT)) %>% as.duration()/dminutes(1)

#number of NAs in BedToDispo
is.na(strep$BedToDispo.mins) %>% table(strep$MnemonicName)
```

```{r}
# Compute Bed Assigned to Discharge
strep$BedToDischarge.mins<- interval(ymd_hms(strep$BedAssign_DT), ymd_hms(strep$Discharge_DateTime)) %>% as.duration()/dminutes(1)

# number of NAs in BedToDischarge
is.na(strep$BedToDischarge.mins) %>% table(strep$MnemonicName)
```


```{r}
#Compute AdmitToDischarge
strep$AdmitToDischarge.mins<- interval(ymd_hms(strep$Admit_DateTime), ymd_hms(strep$Discharge_DateTime)) %>% as.duration()/dminutes(1)

#number of NAs in AdmitToDischarge
is.na(strep$AdmitToDischarge.mins) %>% table(strep$MnemonicName)
```

```{r}
#export strep with dur vars
saveRDS(strep, file = "dfs/df.strep.dur.Rda")
```


##Compare populations for homogeneity

Sex Comparison - No significant difference
```{r}
table(strep$Sex, strep$MnemonicName)
chisq.test(strep$Sex, strep$MnemonicName, correct=FALSE)
```

Age Comparison - No significant difference
```{r}
#Table
do.call("rbind", describeBy(strep$Age, strep$MnemonicName))

#Shapiro
with(strep, shapiro.test(Age[MnemonicName == "Strep A (POCT)"]))
with(strep, shapiro.test(Age[MnemonicName == "Rapid Strep A Antigen"]))

#Wilcox
wilcox.test(Age ~ MnemonicName, data = strep, alternative = "two.sided", var.equal = FALSE, conf.int = TRUE, conf.level = 0.95)
```

##Compare Durations

Admit To Discharge - Strep POCT is signficantly longer
```{r}
do.call("rbind", describeBy(strep$AdmitToDischarge.mins, strep$MnemonicName))

#Shapiro
with(strep, shapiro.test(AdmitToDischarge.mins[MnemonicName == "Strep A (POCT)"]))
with(strep, shapiro.test(AdmitToDischarge.mins[MnemonicName == "Rapid Strep A Antigen"]))

#Wilcox
wilcox.test(AdmitToDischarge.mins ~ MnemonicName, data = strep, alternative = "two.sided", var.equal = FALSE, conf.int = TRUE, conf.level = 0.95)
```

Bed to Dispo - Strep POCT is signficantly shorter
```{r}
do.call("rbind", describeBy(strep$BedToDispo.mins, strep$MnemonicName))

#Shapiro
with(strep, shapiro.test(BedToDispo.mins[MnemonicName == "Strep A (POCT)"]))
with(strep, shapiro.test(BedToDispo.mins[MnemonicName == "Rapid Strep A Antigen"]))

#Wilcox
wilcox.test(BedToDispo.mins ~ MnemonicName, data = strep, alternative = "two.sided", var.equal = FALSE, conf.int = TRUE, conf.level = 0.95)
```

Door to Bed
```{r}
do.call("rbind", describeBy(strep$DoorToBed.mins, strep$MnemonicName))

#Shapiro
with(strep, shapiro.test(DoorToBed.mins[MnemonicName == "Strep A (POCT)"]))
with(strep, shapiro.test(DoorToBed.mins[MnemonicName == "Rapid Strep A Antigen"]))

#Wilcox
wilcox.test(DoorToBed.mins ~ MnemonicName, data = strep, alternative = "two.sided", var.equal = FALSE, conf.int = TRUE, conf.level = 0.95)
```

Bed to Discharge - Strep POCT is signficantly shorter
```{r}
do.call("rbind", describeBy(strep$BedToDischarge.mins, strep$MnemonicName))

#Shapiro
with(strep, shapiro.test(BedToDischarge.mins[MnemonicName == "Strep A (POCT)"]))
with(strep, shapiro.test(BedToDischarge.mins[MnemonicName == "Rapid Strep A Antigen"]))

#Wilcox
wilcox.test(BedToDischarge.mins ~ MnemonicName, data = strep, alternative = "two.sided", var.equal = FALSE, conf.int = TRUE, conf.level = 0.95)
```
