==> 9 - RSV Analysis.Rmd <==
---
title: "RSV Analysis"
output: html_notebook
---


```{r}
#Import & libraries
rsv <- readRDS(file = "dfs/df.rsv.Rda") 

levels(rsv$MnemonicName)

#reorder MnemonicName levels
rsv$MnemonicName <- rsv$MnemonicName %>% factor(levels = c("RSV (POCT)","Respiratory PCR Panel"))

library(tidyverse)
library(lubridate)
library(psych)
```

##Compute Time Duration Variables

```{r}
#Compute Bed Assigned to Discharge Decision
rsv$BedToDispo.mins<- interval(ymd_hms(rsv$BedAssign_DT), ymd_hms(rsv$DischargeDecision_DT)) %>% as.duration()/dminutes(1)

#number of NAs in BedToDispo
is.na(rsv$BedToDispo.mins) %>% table(rsv$MnemonicName)
```

```{r}
# Compute Bed Assigned to Discharge
rsv$BedToDischarge.mins<- interval(ymd_hms(rsv$BedAssign_DT), ymd_hms(rsv$Discharge_DateTime)) %>% as.duration()/dminutes(1)

# number of NAs in BedToDischarge
is.na(rsv$BedToDischarge.mins) %>% table(rsv$MnemonicName)
```


```{r}
#Compute AdmitToDischarge
rsv$AdmitToDischarge.mins<- interval(ymd_hms(rsv$Admit_DateTime), ymd_hms(rsv$Discharge_DateTime)) %>% as.duration()/dminutes(1)

#number of NAs in AdmitToDischarge
is.na(rsv$AdmitToDischarge.mins) %>% table(rsv$MnemonicName)
```

```{r}
#export rsv with durs
saveRDS(rsv, file = "dfs/df.rsv.durs.Rda")
```


##Compare populations for homogeneity

Sex Comparison - No significant difference
```{r}
table(rsv$Sex, rsv$MnemonicName)
chisq.test(rsv$Sex, rsv$MnemonicName, correct=FALSE)
```

Age Comparison - No significant difference
```{r}
#Table
do.call("rbind", describeBy(rsv$Age, rsv$MnemonicName))

#Shapiro
with(rsv, shapiro.test(Age[MnemonicName == "RSV (POCT)"]))
with(rsv, shapiro.test(Age[MnemonicName == "Respiratory PCR Panel"]))

#Wilcox
wilcox.test(Age ~ MnemonicName, data = rsv, alternative = "two.sided", var.equal = FALSE, conf.int = TRUE, conf.level = 0.95)
```

##Compare Durations

Admit To Discharge - RSV POCT is signficantly shorter
```{r}
do.call("rbind", describeBy(rsv$AdmitToDischarge.mins, rsv$MnemonicName))

#Shapiro
with(rsv, shapiro.test(AdmitToDischarge.mins[MnemonicName == "RSV (POCT)"]))
with(rsv, shapiro.test(AdmitToDischarge.mins[MnemonicName == "Respiratory PCR Panel"]))

#Wilcox
wilcox.test(AdmitToDischarge.mins ~ MnemonicName, data = rsv, alternative = "two.sided", var.equal = FALSE, conf.int = TRUE, conf.level = 0.95)
```

Bed to Dispo - rsv POCT is signficantly shorter
```{r}
do.call("rbind", describeBy(rsv$BedToDispo.mins, rsv$MnemonicName))

#Shapiro
with(rsv, shapiro.test(BedToDispo.mins[MnemonicName == "RSV (POCT)"]))
with(rsv, shapiro.test(BedToDispo.mins[MnemonicName == "Respiratory PCR Panel"]))

#Wilcox
wilcox.test(BedToDispo.mins ~ MnemonicName, data = rsv, alternative = "two.sided", var.equal = FALSE, conf.int = TRUE, conf.level = 0.95)
```
