==> 10 - Flu Order Patterns Comparison.Rmd <==
---
title: "Flu Order Patterns Comparison"
output: html_notebook
---

```{r}
#Import & libraries
flu.poct <- readRDS(file = "dfs/only.poct.flu.Rda")
flu.lab <- readRDS(file = "dfs/only.lab.pcr.Rda")
track <- readRDS(file = "dfs/track.dedup.Rda")

library(tidyverse)
library(lubridate)
library(psych)
library(ggpubr)
library(reshape)
```


```{r}
#filter out encounters to match date range
track <- track %>% subset(Arrival_DT >= as.Date("2017-10-01"))
```

Count number of encounters per epi_yearweek
```{r}
track$epi_week <- epiweek(track$Arrival_DT)
track$epi_year <- epiyear(track$Arrival_DT)
track$epi_yearweek <- if_else(track$epi_week<10, paste(track$epi_year, "-0", track$epi_week, sep = ""), paste(track$epi_year, track$epi_week, sep = "-"))
track <- track %>% arrange(Arrival_DT)
weekly.track <- track %>% count(epi_yearweek)

ggline(weekly.track, "epi_yearweek", "n")
```

Count number of Flu POCT per epi_yearweek
```{r}
flu.poct$epi_week <- epiweek(flu.poct$Arrival_DT)
flu.poct$epi_year <- epiyear(flu.poct$Arrival_DT)
flu.poct$epi_yearweek <- if_else(flu.poct$epi_week<10, paste(flu.poct$epi_year, "-0", flu.poct$epi_week, sep = ""), paste(flu.poct$epi_year, flu.poct$epi_week, sep = "-"))
flu.poct <- flu.poct %>% arrange(Arrival_DT)
weekly.flu.poct <- flu.poct %>% count(epi_yearweek)

ggline(weekly.flu.poct, "epi_yearweek", "n")
```

Count number of Flu lab per epi_yearweek
```{r}
flu.lab$epi_week <- epiweek(flu.lab$Arrival_DT)
flu.lab$epi_year <- epiyear(flu.lab$Arrival_DT)
flu.lab$epi_yearweek <- if_else(flu.lab$epi_week<10, paste(flu.lab$epi_year, "-0", flu.lab$epi_week, sep = ""), paste(flu.lab$epi_year, flu.lab$epi_week, sep = "-"))
flu.lab <- flu.lab %>% arrange(Arrival_DT)

weekly.flu.lab <- flu.lab %>% count(epi_yearweek)

ggline(weekly.flu.lab, "epi_yearweek", "n")
```

Merge data sets by epi_yearweek
```{r}
#first merge track with flu.lab
weekly.merged<- merge(x = weekly.track,
                      y = weekly.flu.lab,
                      by = "epi_yearweek",
                      all.x = TRUE) %>%
  rename("Encounters" = n.x, "Respiratory PCR Panel Orders" = n.y)

#then merge with flu.poct
weekly.merged <- merge(x = weekly.merged,
                      y = weekly.flu.poct,
                      by = "epi_yearweek",
                      all.x = TRUE) %>%
  rename("POCT Flu Orders" = n)

#force NA to 0
weekly.merged[is.na(weekly.merged)] = 0
```

Subset track by ILI
```{r}
#fever, cough, sore throat, flu, cold, body*ache, chills
track.ili <- track %>% subset(
  grepl("fever",track$RFV_ED, ignore.case=TRUE)| 
  grepl("cough",track$RFV_ED, ignore.case=TRUE)|
  grepl("sore.*throat",track$RFV_ED, ignore.case=TRUE)|
  grepl("throat.*pain",track$RFV_ED, ignore.case=TRUE)|
  grepl("flu",track$RFV_ED, ignore.case=TRUE)|
  grepl("cold",track$RFV_ED, ignore.case=TRUE)|
  grepl("body.*ache",track$RFV_ED, ignore.case=TRUE)|
  grepl("chills",track$RFV_ED, ignore.case=TRUE)
)

weekly.track.ili <- track.ili %>% count(epi_yearweek)

ggline(weekly.track.ili, "epi_yearweek", "n")
```


Create data set with all encounters of track (leftjoin) and added poct and lab flu
```{r}
trackprop <- merge(x = track[,c("FIN", "epi_yearweek", "Arrival_DT", "DOB", "Sex", "RFV_ED")],
                   y = flu.lab[,c("FIN", "MnemonicName")],
                   by = "FIN",
                   all.x = TRUE)
                   
trackprop %>% count(Resp_PCR_Ordered)

trackprop <- merge(x = trackprop,
                   y = flu.poct[,c("FIN", "MnemonicName")],
                   by = "FIN",
                   all.x = TRUE)

trackprop <- trackprop %>% rename("Resp_PCR_Ordered" = MnemonicName.x,
                    "Flu_POCT_Ordereed" = MnemonicName.y)

trackprop <- trackprop %>% rename("Flu_POCT_Ordered" = Flu_POCT_Ordereed)

trackprop$Resp_PCR_Ordered <- ifelse(is.na(trackprop$Resp_PCR_Ordered), 0, 1) %>% #convert Resp PCR to factors
  factor(levels = c(0,1),
         labels = c("Not Ordered", "Ordered"))

trackprop$Flu_POCT_Ordered <- ifelse(is.na(trackprop$Flu_POCT_Ordered), 0, 1) %>% #convert Flu to factors
  factor(levels = c(0,1),
         labels = c("Not Ordered", "Ordered"))
```

Subset that data for ILI only
```{r}
#fever, cough, sore throat, flu, cold, body*ache, chills
trackprop.ili <- trackprop %>% subset(
  grepl("fever",trackprop$RFV_ED, ignore.case=TRUE)| 
  grepl("cough",trackprop$RFV_ED, ignore.case=TRUE)|
  grepl("sore.*throat",trackprop$RFV_ED, ignore.case=TRUE)|
  grepl("throat.*pain",trackprop$RFV_ED, ignore.case=TRUE)|
  grepl("flu",trackprop$RFV_ED, ignore.case=TRUE)|
  grepl("cold",trackprop$RFV_ED, ignore.case=TRUE)|
  grepl("body.*ache",trackprop$RFV_ED, ignore.case=TRUE)|
  grepl("chills",trackprop$RFV_ED, ignore.case=TRUE)
)
```

##For all encounters

Resp PCR before 10/1/18, POCT after 10/1/18
```{r}
trackpropA <- trackprop %>% subset(Arrival_DT <= as.Date("2018-10-01"))
trackpropB <- trackprop %>% subset(Arrival_DT >= as.Date("2018-10-01"))

trackprop %>% 
  subset(Arrival_DT <= as.Date("2018-10-01")) %>% 
  count(Resp_PCR_Ordered)

trackprop %>% subset(Arrival_DT >= as.Date("2018-10-01")) %>%
  count(Flu_POCT_Ordered)

RespPCRfreq <- table(trackpropA$Resp_PCR_Ordered) %>% as.data.frame() 
names(RespPCRfreq)[1] = "Status"
names(RespPCRfreq)[2] = "Resp_PCR"

FluPOCTfreq <- table(trackpropB$Flu_POCT_Ordered) %>% as.data.frame() 
names(FluPOCTfreq)[1] = "Status"
names(FluPOCTfreq)[2] = "Flu_POCT"

freq.all.table <- merge(x = FluPOCTfreq, 
                        y = RespPCRfreq,
                        by = "Status") %>% t()

```
