==> 11- Alere Data Import.Rmd <==
---
title: "Alere Data Import"
output: html_notebook
---

```{r}
#Import & libraries
alere <- read_excel("~/Library/Mobile Documents/com~apple~CloudDocs/Documents/EM Research/Data Analysis/RVPv3.4/RawData/Alere_RALS_2019-02-11.xls", 
                    col_types = c("date", "text", "text", 
                                  "text", "text", "text", "text", "text", 
                                  "text", "text", "text", "text", "text", 
                                  "text", "text", "text", "text", "text"))

library(tidyverse)
library(lubridate)
library(psych)
```

```{r}
#subset, removing QCs
alere<- alere %>% subset(SampleType == "Patient")

#Remove vars that don't matter
alere <- alere %>% select(Alere_DateTime, PatientID, TestType, StrepA, FluA, FluB, RSV, Flags)


#remove unusable encounters
alere <- alere %>% filter(!grepl("test",PatientID))
alere <- alere %>% filter(!grepl("^0*$",PatientID))
alere <- alere %>% filter(!grepl("^1*$",PatientID))
alere <- alere %>% filter(!grepl("^2*$",PatientID))
alere <- alere %>% filter(!grepl("^3*$",PatientID))
alere <- alere %>% filter(!grepl("^4*$",PatientID))
alere <- alere %>% filter(!grepl("^5*$",PatientID))
alere <- alere %>% filter(!grepl("^6*$",PatientID))
alere <- alere %>% filter(!grepl("^7*$",PatientID))

alere %>% filter(grepl("\\d",PatientID))

#Patient ID Character Distributions - Most have 12
nchar(alere$PatientID) %>% factor() %>% summary()

#Create Possible FIN row, keeping those w/ only 12 digits
alere$FIN <- 
  ifelse(nchar(alere$PatientID) == 12 & grepl("\\d",alere$PatientID),
         alere$PatientID,
         NA)

#export
saveRDS(alere, file = "dfs/alere.all.Rda")
```
