==> 12-Alere & Flue, Merge & Analysis.Rmd <==
---
title: "Alere & Flue Merge & Analysis"
output: html_notebook
---


```{r}
flu <- readRDS(file = "dfs/df.flu.Rda")
alere <- readRDS(file = "dfs/alere.all.Rda")

library(tidyverse)
library(lubridate)
library(psych)
```

Filter out to include just flu-relevant for alere
```{r}
alere <-
  alere %>% subset(TestType == "Flu A/B")

#Create overall column to combine FluA/B
alere$FluOverall <-
  case_when(alere$FluA == "positive" | alere$FluB == "positive" ~ "positive",
           alere$FluA == "negative" | alere$FluB == "negative" ~ "negative",
           alere$FluA == "invalid" | alere$FluB == "invalid" ~ "invalid"
           ) %>% 
  factor(levels = c("positive", "negative", "invalid"))



#remove now unnecessary columns, change order
alere <- 
  alere %>% select(Alere_DateTime, FIN, FluA, FluB, FluOverall, Flags)

alere %>% count(FluOverall)

#export
saveRDS(alere, file = "dfs/alere.clean.Rda")
```

```{r}
# create column for inconclusive test
alere$Inconclusive <- case_when(grepl(".*Inconclusive.*",alere$Flags, ignore.case=TRUE) ~ 1,
                                TRUE ~ 0) %>%
  factor(levels = c(0,1),
         labels = c("Not Inconclusive", "Inconclusive"))

alere %>% count(Inconclusive)
```


Show Duplicates
```{r}
alere.dup <- alere[duplicated(alere$FIN)|duplicated(alere$FIN, fromLast = TRUE),] %>% #show encounters with duplicate FIN
  arrange(FIN, Inconclusive, FluOverall, Alere_DateTime) #sort so inconclusive before postive before negative before invalid

alere.dup %>% count(FluOverall)
alere.dup %>% count(Inconclusive)
```

Remove duplicates
```{r}
#apply above duplicate sort to full set
alere.dedup <- alere %>% arrange(FIN, Inconclusive, FluOverall, Alere_DateTime)

#remove duplicates
alere.dedup <- alere.dedup %>% distinct(FIN, .keep_all = TRUE) #remove duplicates (removes 2nd+ row)

#re sort
alere.dedup <- alere.dedup %>% arrange(Alere_DateTime)
  
alere.dedup %>% count(FluOverall)
alere.dedup %>% count(Inconclusive)

table(alere.dedup$FluOverall, alere.dedup$Inconclusive)

#export
saveRDS(alere.dedup, file = "dfs/alere.dedup.Rda")
```

Merge Flu+Tracking with Alere
```{r}
alere.flu <- merge(x = flu,
                   y = alere.dedup,
                   by = "FIN",
                   all.x = TRUE)

alere.flu %>% filter(MnemonicName == "Influenza (POCT)") %>% count(FluOverall)

table(alere.flu$MnemonicName, alere.flu$FluOverall)
alere.flu %>% count(MnemonicName)
```

Merge track with alere
```{r}
track <- readRDS(file = "dfs/track.dedup.Rda")
alere.track <- merge(x = track,
                   y = alere.dedup,
                   by = "FIN",
                   all = FALSE) 

alere.dedup %>% count(FluOverall)
alere.track %>% count(FluOverall)
              
```

Add Duration Vars
```{r}
#Compute AdmitToDischarge
alere.flu$AdmitToDischarge.mins<- interval(ymd_hms(alere.flu$Admit_DateTime), ymd_hms(alere.flu$Discharge_DateTime)) %>% as.duration()/dminutes(1)

#Compute Bed Assigned to Discharge Decision
alere.flu$BedToDispo.mins<- interval(ymd_hms(alere.flu$BedAssign_DT), ymd_hms(alere.flu$DischargeDecision_DT)) %>% as.duration()/dminutes(1)
```

Compare pos vs neg - LOS
```{r}
#remove invalid from group
alere.flu.noinvalid <- alere.flu %>% filter(FluOverall != "invalid")

#stats
do.call("rbind", describeBy(alere.flu.noinvalid$AdmitToDischarge.mins, alere.flu.noinvalid$FluOverall))

#comparison
wilcox.test(AdmitToDischarge.mins ~ FluOverall, data = alere.flu.noinvalid, alternative = "two.sided", var.equal = FALSE, conf.int = TRUE, conf.level = 0.95)


#Export
saveRDS(alere.flu, file = "dfs/alere.flu.Rda")
saveRDS(alere.flu.noinvalid, file = "dfs/alere.flu.noinvalid.Rda")

```

Compare pos vs neg - BedToDispo
```{r}
#stats
do.call("rbind", describeBy(alere.flu.noinvalid$BedToDispo.mins, alere.flu.noinvalid$FluOverall))

#comparison
wilcox.test(BedToDispo.mins ~ FluOverall, data = alere.flu.noinvalid, alternative = "two.sided", var.equal = FALSE, conf.int = TRUE, conf.level = 0.95)

```
