==> 15 - Redo 13 (Flu and LOSCXR) with New Full Set.Rmd <==
---
title: "15 - Redo Flu and LOS & CXR Analysis"
output: html_notebook
---

##Setup
```{r}
flu <- readRDS(file = "dfs/2ndx/df.flu.beforeafter2.dedup.Rda")
library(tidyverse)
library(lubridate)
library(psych)
```

##Compute Durations

```{r}
#Compute Bed Assigned to Discharge Decision
flu$BedToDispo.mins<- interval(ymd_hms(flu$BedAssign_DT), ymd_hms(flu$DischargeDecision_DT)) %>% as.duration()/dminutes(1)

#number of NAs in BedToDispo
is.na(flu$BedToDispo.mins) %>% table(flu$MnemonicName)
```

```{r}
#Compute AdmitToDischarge
flu$AdmitToDischarge.mins<- interval(ymd_hms(flu$Checkin_DT), ymd_hms(flu$Checkout_DT)) %>% as.duration()/dminutes(1)

#number of NAs in AdmitToDischarge
is.na(flu$AdmitToDischarge.mins) %>% table(flu$MnemonicName)
```

##Traditonal Flu Outlier Removal (2294 to 2281 = 13 removed)
```{r}
#Remove predefined outliers (removes values LOS <25 min); 
alere.flu <- flu %>% filter(!(AdmitToDischarge.mins < 15))
```

Subset into Positive & Negative
```{r}
alere.flu %>% group_by(MnemonicName2) %>% count(FluOverall)

alere.flu.pos <- alere.flu %>% filter(
  MnemonicName2 == "Respiratory PCR Panel"|
  FluOverall == "positive")


alere.flu.neg <- alere.flu %>% filter(
  MnemonicName2 == "Respiratory PCR Panel"|
  FluOverall == "negative")
```
##All Encounter (All) LOS
Shapiro Test - failed
```{r}
with(alere.flu, shapiro.test(AdmitToDischarge.mins[MnemonicName == "Influenza (POCT)"]))
with(alere.flu, shapiro.test(AdmitToDischarge.mins[MnemonicName == "Respiratory PCR Panel"]))
```

Wilcoxon test - significant
```{r}
do.call("rbind", describeBy(alere.flu$AdmitToDischarge.mins, alere.flu$MnemonicName)) %>% select(n, mean, median, min, max)
wilcox.test(AdmitToDischarge.mins ~ MnemonicName, data = alere.flu, alternative = "two.sided", var.equal = FALSE, conf.int = TRUE, conf.level = 0.95)
```

Bootstrapped Median CI
```{r}
library(rcompanion)
groupwiseMedian(data=alere.flu, AdmitToDischarge.mins ~ MnemonicName, bca = FALSE, percentile = TRUE, R = 1000)
```



##All Encounter (Positive) LOS
Shapiro Test - failed
```{r}
with(alere.flu.pos, shapiro.test(AdmitToDischarge.mins[MnemonicName == "Influenza (POCT)"]))
with(alere.flu.pos, shapiro.test(AdmitToDischarge.mins[MnemonicName == "Respiratory PCR Panel"]))
```

Wilcoxon test - significant
```{r}
do.call("rbind", describeBy(alere.flu.pos$AdmitToDischarge.mins, alere.flu.pos$MnemonicName)) %>% select(n, mean, median, min, max)
wilcox.test(AdmitToDischarge.mins ~ MnemonicName, data = alere.flu.pos, alternative = "two.sided", var.equal = FALSE, conf.int = TRUE, conf.level = 0.95)
```

Bootstrapped Median CI
```{r}
library(rcompanion)
groupwiseMedian(data=alere.flu.pos, AdmitToDischarge.mins ~ MnemonicName, bca = FALSE, percentile = TRUE, R = 1000)
```

##All Encounter (Negative) LOS
Shapiro Test - failed
```{r}
with(alere.flu.neg, shapiro.test(AdmitToDischarge.mins[MnemonicName == "Influenza (POCT)"]))
with(alere.flu.neg, shapiro.test(AdmitToDischarge.mins[MnemonicName == "Respiratory PCR Panel"]))
```

Wilcoxon test - significant
```{r}
do.call("rbind", describeBy(alere.flu.neg$AdmitToDischarge.mins, alere.flu.neg$MnemonicName)) %>% select(n, mean, median, min, max)
wilcox.test(AdmitToDischarge.mins ~ MnemonicName, data = alere.flu.neg, alternative = "two.sided", var.equal = FALSE, conf.int = TRUE, conf.level = 0.95)
```

Bootstrapped Median CI
```{r}
library(rcompanion)
groupwiseMedian(data=alere.flu.neg, AdmitToDischarge.mins ~ MnemonicName, bca = FALSE, percentile = TRUE, R = 1000)
```

##All Encounter (Positive) BedToDispo
Shapiro Test - failed
```{r}
with(alere.flu.pos, shapiro.test(BedToDispo.mins[MnemonicName == "Influenza (POCT)"]))
with(alere.flu.pos, shapiro.test(BedToDispo.mins[MnemonicName == "Respiratory PCR Panel"]))
```

Wilcoxon test - significant
```{r}
do.call("rbind", describeBy(alere.flu.pos$BedToDispo.mins, alere.flu.pos$MnemonicName)) %>% select(n, mean, median, min, max)
wilcox.test(BedToDispo.mins ~ MnemonicName, data = alere.flu.pos, alternative = "two.sided", var.equal = FALSE, conf.int = TRUE, conf.level = 0.95)
```

Bootstrapped Median CI
```{r}
library(rcompanion)
groupwiseMedian(data=filter(alere.flu.pos, !is.na(BedToDispo.mins)), BedToDispo.mins ~ MnemonicName, bca = FALSE, percentile = TRUE, R = 1000)
```

##All Encounter (Negative) BedToDispo

Shapiro Test - failed
```{r}
with(alere.flu.neg, shapiro.test(BedToDispo.mins[MnemonicName == "Influenza (POCT)"]))
with(alere.flu.neg, shapiro.test(BedToDispo.mins[MnemonicName == "Respiratory PCR Panel"]))
```

Wilcoxon test - significant
```{r}
do.call("rbind", describeBy(alere.flu.neg$BedToDispo.mins, alere.flu.neg$MnemonicName)) %>% select(n, mean, median, min, max)
wilcox.test(BedToDispo.mins ~ MnemonicName, data = alere.flu.neg, alternative = "two.sided", var.equal = FALSE, conf.int = TRUE, conf.level = 0.95)
```

Bootstrapped Median CI
```{r}
library(rcompanion)
groupwiseMedian(data=filter(alere.flu.neg, !is.na(BedToDispo.mins)), BedToDispo.mins ~ MnemonicName, bca = FALSE, percentile = TRUE, R = 1000)
```



##All Encounter Flu CXR Analysis

All
```{r}
table(alere.flu$MnemonicName, alere.flu$XR_Completed)
chisq.test(alere.flu$MnemonicName, alere.flu$XR_Completed, correct = FALSE)
```


Positive
```{r}
table(alere.flu.pos$MnemonicName, alere.flu.pos$XR_Completed)
chisq.test(alere.flu.pos$MnemonicName, alere.flu.pos$XR_Completed, correct = FALSE)
```

Negative
```{r}
table(alere.flu.neg$MnemonicName, alere.flu.neg$XR_Completed)
chisq.test(alere.flu.neg$MnemonicName, alere.flu.neg$XR_Completed, correct = FALSE)
```


##Traditional Flu Bed to Dispo Analysis

Bed to Dispo Stats
```{r}
do.call("rbind", describeBy(flu$BedToDispo.mins, flu$MnemonicName))
```

Shapiro Test - failed
```{r}
with(flu, shapiro.test(BedToDispo.mins[MnemonicName == "Influenza (POCT)"]))
with(flu, shapiro.test(BedToDispo.mins[MnemonicName == "Respiratory PCR Panel"]))
```

Wilcoxon test - significant
```{r}
wilcox.test(BedToDispo.mins ~ MnemonicName, data = flu, alternative = "two.sided", var.equal = FALSE, conf.int = TRUE, conf.level = 0.95)
```


##Modified ILI Flu Setup
```{r}
#fever, cough, sore throat, flu, cold, body*ache, chills
alere.flu.ili <-alere.flu %>% subset(
  grepl("fever",alere.flu$RFV, ignore.case=TRUE)| 
  grepl("cough",alere.flu$RFV, ignore.case=TRUE)|
  grepl("sore.*throat",alere.flu$RFV, ignore.case=TRUE)|
  grepl("throat.*pain",alere.flu$RFV, ignore.case=TRUE)|
  grepl("flu",alere.flu$RFV, ignore.case=TRUE)|
  grepl("cold",alere.flu$RFV, ignore.case=TRUE)|
  grepl("body.*ache",alere.flu$RFV, ignore.case=TRUE)|
  grepl("chills",alere.flu$RFV, ignore.case=TRUE)
)

do.call("rbind", describeBy(alere.flu.ili$AdmitToDischarge.mins, alere.flu.ili$MnemonicName))
```
Positive
```{r}
#fever, cough, sore throat, flu, cold, body*ache, chills
alere.flu.ili.pos <-alere.flu.pos %>% subset(
  grepl("fever",alere.flu.pos$RFV, ignore.case=TRUE)| 
  grepl("cough",alere.flu.pos$RFV, ignore.case=TRUE)|
  grepl("sore.*throat",alere.flu.pos$RFV, ignore.case=TRUE)|
  grepl("throat.*pain",alere.flu.pos$RFV, ignore.case=TRUE)|
  grepl("flu",alere.flu.pos$RFV, ignore.case=TRUE)|
  grepl("cold",alere.flu.pos$RFV, ignore.case=TRUE)|
  grepl("body.*ache",alere.flu.pos$RFV, ignore.case=TRUE)|
  grepl("chills",alere.flu.pos$RFV, ignore.case=TRUE)
)

do.call("rbind", describeBy(alere.flu.ili.pos$AdmitToDischarge.mins, alere.flu.ili.pos$MnemonicName))
```

Negative
```{r}
#fever, cough, sore throat, flu, cold, body*ache, chills
alere.flu.ili.neg <-alere.flu.neg %>% subset(
  grepl("fever",alere.flu.neg$RFV, ignore.case=TRUE)| 
  grepl("cough",alere.flu.neg$RFV, ignore.case=TRUE)|
  grepl("sore.*throat",alere.flu.neg$RFV, ignore.case=TRUE)|
  grepl("throat.*pain",alere.flu.neg$RFV, ignore.case=TRUE)|
  grepl("flu",alere.flu.neg$RFV, ignore.case=TRUE)|
  grepl("cold",alere.flu.neg$RFV, ignore.case=TRUE)|
  grepl("body.*ache",alere.flu.neg$RFV, ignore.case=TRUE)|
  grepl("chills",alere.flu.neg$RFV, ignore.case=TRUE)
)

do.call("rbind", describeBy(alere.flu.ili.neg$AdmitToDischarge.mins, alere.flu.ili.neg$MnemonicName))
```

##Modified ILI Flu Homegeneity of Groups
###Sex Comparison
```{r}
alere.flu.ili$MnemonicName %>% ftable(alere.flu.ili$Sex)
chisq.test(alere.flu.ili$MnemonicName, alere.flu.ili$Sex, correct=FALSE)
```

###Age Comparison

```{r}
do.call("rbind", describeBy(alere.flu.ili$Age, alere.flu.ili$MnemonicName))
wilcox.test(Age ~ MnemonicName, data = alere.flu.ili, alternative = "two.sided", var.equal = FALSE, conf.int = TRUE, conf.level = 0.95)
```


##Modified ILI Flu (All) LOS
Shapiro Test - failed
```{r}
with(flu.ili, shapiro.test(AdmitToDischarge.mins[MnemonicName == "Influenza (POCT)"]))
with(flu.ili, shapiro.test(AdmitToDischarge.mins[MnemonicName == "Respiratory PCR Panel"]))
```

Wilcoxon test - significant
```{r}
do.call("rbind", describeBy(flu.ili$AdmitToDischarge.mins, flu.ili$MnemonicName)) %>% select(n, mean, median, min, max)
wilcox.test(AdmitToDischarge.mins ~ MnemonicName, data = flu.ili, alternative = "two.sided", var.equal = FALSE, conf.int = TRUE, conf.level = 0.95)
```

Bootstrapped Median CI
```{r}
library(rcompanion)
groupwiseMedian(data=flu.ili, AdmitToDischarge.mins ~ MnemonicName, bca = FALSE, percentile = TRUE, R = 1000)
```

##Modified ILI (Positive) LOS
Shapiro Test - failed
```{r}
with(alere.flu.ili.pos, shapiro.test(AdmitToDischarge.mins[MnemonicName == "Influenza (POCT)"]))
with(alere.flu.ili.pos, shapiro.test(AdmitToDischarge.mins[MnemonicName == "Respiratory PCR Panel"]))
```

Wilcoxon test - significant
```{r}
do.call("rbind", describeBy(alere.flu.ili.pos$AdmitToDischarge.mins, alere.flu.ili.pos$MnemonicName)) %>% select(n, mean, median, min, max)
wilcox.test(AdmitToDischarge.mins ~ MnemonicName, data = alere.flu.ili.pos, alternative = "two.sided", var.equal = FALSE, conf.int = TRUE, conf.level = 0.95)
```

Bootstrapped Median CI
```{r}
library(rcompanion)
groupwiseMedian(data=alere.flu.ili.pos, AdmitToDischarge.mins ~ MnemonicName, bca = FALSE, percentile = TRUE, R = 1000)
```

##Modified ILI (Negative) LOS
Shapiro Test - failed
```{r}
with(alere.flu.ili.neg, shapiro.test(AdmitToDischarge.mins[MnemonicName == "Influenza (POCT)"]))
with(alere.flu.ili.neg, shapiro.test(AdmitToDischarge.mins[MnemonicName == "Respiratory PCR Panel"]))
```

Wilcoxon test - significant
```{r}
do.call("rbind", describeBy(alere.flu.ili.neg$AdmitToDischarge.mins, alere.flu.ili.neg$MnemonicName)) %>% select(n, mean, median, min, max)
wilcox.test(AdmitToDischarge.mins ~ MnemonicName, data = alere.flu.ili.neg, alternative = "two.sided", var.equal = FALSE, conf.int = TRUE, conf.level = 0.95)
```

Bootstrapped Median CI
```{r}
library(rcompanion)
groupwiseMedian(data=alere.flu.ili.neg, AdmitToDischarge.mins ~ MnemonicName, bca = FALSE, percentile = TRUE, R = 1000)
```

##Modified ILI (All) BedToDispo
Shapiro Test - failed
```{r}
with(alere.flu.ili, shapiro.test(BedToDispo.mins[MnemonicName == "Influenza (POCT)"]))
with(alere.flu.ili, shapiro.test(BedToDispo.mins[MnemonicName == "Respiratory PCR Panel"]))
```

Wilcoxon test - significant
```{r}
do.call("rbind", describeBy(alere.flu.ili$BedToDispo.mins, alere.flu.ili$MnemonicName)) %>% select(n, mean, median, min, max)
wilcox.test(BedToDispo.mins ~ MnemonicName, data = alere.flu.ili, alternative = "two.sided", var.equal = FALSE, conf.int = TRUE, conf.level = 0.95)
```

Bootstrapped Median CI
```{r}
library(rcompanion)
groupwiseMedian(data=filter(alere.flu.ili, !is.na(BedToDispo.mins)), BedToDispo.mins ~ MnemonicName, bca = FALSE, percentile = TRUE, R = 1000)
```

##Modified ILI (Positive) BedToDispo
Shapiro Test - failed
```{r}
with(alere.flu.ili.pos, shapiro.test(BedToDispo.mins[MnemonicName == "Influenza (POCT)"]))
with(alere.flu.ili.pos, shapiro.test(BedToDispo.mins[MnemonicName == "Respiratory PCR Panel"]))
```

Wilcoxon test - significant
```{r}
do.call("rbind", describeBy(alere.flu.ili.pos$BedToDispo.mins, alere.flu.ili.pos$MnemonicName)) %>% select(n, mean, median, min, max)
wilcox.test(BedToDispo.mins ~ MnemonicName, data = alere.flu.ili.pos, alternative = "two.sided", var.equal = FALSE, conf.int = TRUE, conf.level = 0.95)
```

Bootstrapped Median CI
```{r}
library(rcompanion)
groupwiseMedian(data=filter(alere.flu.ili.pos, !is.na(BedToDispo.mins)), BedToDispo.mins ~ MnemonicName, bca = FALSE, percentile = TRUE, R = 1000)
```

##Modified ILI (Negative) BedToDispo

Shapiro Test - failed
```{r}
with(alere.flu.ili.neg, shapiro.test(BedToDispo.mins[MnemonicName == "Influenza (POCT)"]))
with(alere.flu.ili.neg, shapiro.test(BedToDispo.mins[MnemonicName == "Respiratory PCR Panel"]))
```

Wilcoxon test - significant
```{r}
do.call("rbind", describeBy(alere.flu.ili.neg$BedToDispo.mins, alere.flu.ili.neg$MnemonicName)) %>% select(n, mean, median, min, max)
wilcox.test(BedToDispo.mins ~ MnemonicName, data = alere.flu.ili.neg, alternative = "two.sided", var.equal = FALSE, conf.int = TRUE, conf.level = 0.95)
```

Bootstrapped Median CI
```{r}
library(rcompanion)
groupwiseMedian(data=filter(alere.flu.ili.neg, !is.na(BedToDispo.mins)), BedToDispo.mins ~ MnemonicName, bca = FALSE, percentile = TRUE, R = 1000)
```



##Modified ILI Flu CXR Analysis

All
```{r}
table(alere.flu.ili$MnemonicName, alere.flu.ili$XR_Completed)
chisq.test(alere.flu.ili$MnemonicName, alere.flu.ili$XR_Completed, correct = FALSE)
```


Positive
```{r}
table(alere.flu.ili.pos$MnemonicName, alere.flu.ili.pos$XR_Completed)
chisq.test(alere.flu.ili.pos$MnemonicName, alere.flu.ili.pos$XR_Completed, correct = FALSE)
```

Negative
```{r}
table(alere.flu.ili.neg$MnemonicName, alere.flu.ili.neg$XR_Completed)
chisq.test(alere.flu.ili.neg$MnemonicName, alere.flu.ili.neg$XR_Completed, correct = FALSE)
```
