==> 19 - Full Datasets Analysis Pt 1.Rmd <==
---
title: "19 - Full Datasets Analysis Pt 1 - Traditional Flu LOS"
output: html_notebook
---

#Setup
```{r}
flu <- readRDS(file = "dfs/alere.flu.pcr.Rda")

library(tidyverse)
library(lubridate)
library(psych)
```

#Split into positive and negatives
```{r}
table(flu$MnemonicName, flu$FluOverall)

flu.pos <- flu %>% filter(
  FluOverall == "positive")

flu.neg <- flu %>% filter(
  FluOverall == "negative")

```

#All Flu
##Outlier Removal
```{r}
#Remove predefined outliers (removes values LOS <20 min); 
flu2 <- flu %>% filter(!(AdmitToDischarge.mins < 20))
```


##All Flu Sex Comparison
```{r}
flu$Sex %>% ftable(flu$MnemonicName)
chisq.test(flu$MnemonicName, flu$Sex, correct=FALSE)
```

##All Flu Age Comparison

```{r}
do.call("rbind", describeBy(flu$Age, flu$MnemonicName))
wilcox.test(Age ~ MnemonicName, data = flu, alternative = "two.sided", var.equal = FALSE, conf.int = TRUE, conf.level = 0.95)
```

##All Flu LOS Analysis
Shapiro Test - failed
```{r}
with(flu2, shapiro.test(AdmitToDischarge.mins[MnemonicName == "Influenza (POCT)"]))
with(flu2, shapiro.test(AdmitToDischarge.mins[MnemonicName == "Respiratory PCR Panel"]))
```

Wilcoxon test - significant
```{r}
do.call("rbind", describeBy(flu2$AdmitToDischarge.mins, flu2$MnemonicName)) %>% select(n, mean, median, min, max)
wilcox.test(AdmitToDischarge.mins ~ MnemonicName, data = flu2, alternative = "two.sided", var.equal = FALSE, conf.int = TRUE, conf.level = 0.95)
```

Bootstrapped Median CI
```{r}
library(rcompanion)
groupwiseMedian(data=flu2, AdmitToDischarge.mins ~ MnemonicName, bca = FALSE, percentile = TRUE, R = 1000)
```

##All Flu Bed to Dispo Analysis

Shapiro Test - failed
```{r}
with(flu2, shapiro.test(BedToDispo.mins[MnemonicName == "Influenza (POCT)"]))
with(flu2, shapiro.test(BedToDispo.mins[MnemonicName == "Respiratory PCR Panel"]))
```

Wilcoxon test - significant
```{r}
do.call("rbind", describeBy(flu2$BedToDispo.mins, flu2$MnemonicName)) %>% select(n, mean, median, min, max)
wilcox.test(BedToDispo.mins ~ MnemonicName, data = flu2, alternative = "two.sided", var.equal = FALSE, conf.int = TRUE, conf.level = 0.95)
```

Bootstrapped Median CI
```{r}
library(rcompanion)
flu2.btd <- flu2 %>% filter(!is.na(BedToDispo.mins))
groupwiseMedian(data=flu2.btd, BedToDispo.mins ~ MnemonicName, bca = FALSE, percentile = TRUE, R = 1000)
```



#Pos Flu
##Outlier Removal
```{r}
#Remove predefined outliers (removes values LOS <20 min); 
flu.pos2 <- flu.pos %>% filter(!(AdmitToDischarge.mins < 20))
```


##Pos Flu Sex Comparison
```{r}
flu.pos2$Sex %>% ftable(flu.pos2$MnemonicName)
chisq.test(flu.pos2$MnemonicName, flu.pos2$Sex, correct=FALSE)
```

##Pos Flu Age Comparison

```{r}
do.call("rbind", describeBy(flu.pos2$Age, flu.pos2$MnemonicName))
wilcox.test(Age ~ MnemonicName, data = flu.pos2, alternative = "two.sided", var.equal = FALSE, conf.int = TRUE, conf.level = 0.95)
```

##Pos Flu LOS Analysis
Shapiro Test - failed
```{r}
with(flu.pos2, shapiro.test(AdmitToDischarge.mins[MnemonicName == "Influenza (POCT)"]))
with(flu.pos2, shapiro.test(AdmitToDischarge.mins[MnemonicName == "Respiratory PCR Panel"]))
```

Wilcoxon test - 
```{r}
do.call("rbind", describeBy(flu.pos2$AdmitToDischarge.mins, flu.pos2$MnemonicName)) %>% select(n, mean, median, min, max)
wilcox.test(AdmitToDischarge.mins ~ MnemonicName, data = flu.pos2, alternative = "two.sided", var.equal = FALSE, conf.int = TRUE, conf.level = 0.95)
```

Bootstrapped Median CI
```{r}
library(rcompanion)
groupwiseMedian(data=flu.pos2, AdmitToDischarge.mins ~ MnemonicName, bca = FALSE, percentile = TRUE, R = 1000)
```

##Pos Flu Bed to Dispo Analysis

Shapiro Test - failed
```{r}
with(flu.pos2, shapiro.test(BedToDispo.mins[MnemonicName == "Influenza (POCT)"]))
with(flu.pos2, shapiro.test(BedToDispo.mins[MnemonicName == "Respiratory PCR Panel"]))
```

Wilcoxon test - significant
```{r}
do.call("rbind", describeBy(flu.pos2$BedToDispo.mins, flu.pos2$MnemonicName)) %>% select(n, mean, median, min, max)
wilcox.test(BedToDispo.mins ~ MnemonicName, data = flu.pos2, alternative = "two.sided", var.equal = FALSE, conf.int = TRUE, conf.level = 0.95)
```

Bootstrapped Median CI
```{r}
library(rcompanion)
flu.pos2.btd <- flu.pos2 %>% filter(!is.na(BedToDispo.mins))
groupwiseMedian(data=flu.pos2.btd, BedToDispo.mins ~ MnemonicName, bca = FALSE, percentile = TRUE, R = 1000)
```




#Neg Flu
##Outlier Removal
```{r}
#Remove predefined outliers (removes values LOS <20 min); 
flu.neg2 <- flu.neg %>% filter(!(AdmitToDischarge.mins < 20))
```


##Neg Flu Sex Comparison
```{r}
flu.neg2$Sex %>% ftable(flu.neg2$MnemonicName)
chisq.test(flu.neg2$MnemonicName, flu.neg2$Sex, correct=FALSE)
```

##Neg Flu Age Comparison

```{r}
do.call("rbind", describeBy(flu.neg2$Age, flu.neg2$MnemonicName))
wilcox.test(Age ~ MnemonicName, data = flu.neg2, alternative = "two.sided", var.equal = FALSE, conf.int = TRUE, conf.level = 0.95)
```

##Neg Flu LOS Analysis
Shapiro Test - failed
```{r}
with(flu.neg2, shapiro.test(AdmitToDischarge.mins[MnemonicName == "Influenza (POCT)"]))
with(flu.neg2, shapiro.test(AdmitToDischarge.mins[MnemonicName == "Respiratory PCR Panel"]))
```

Wilcoxon test - 
```{r}
do.call("rbind", describeBy(flu.neg2$AdmitToDischarge.mins, flu.neg2$MnemonicName)) %>% select(n, mean, median, min, max)
wilcox.test(AdmitToDischarge.mins ~ MnemonicName, data = flu.neg2, alternative = "two.sided", var.equal = FALSE, conf.int = TRUE, conf.level = 0.95)
```

Bootstrapped Median CI
```{r}
library(rcompanion)
groupwiseMedian(data=flu.neg2, AdmitToDischarge.mins ~ MnemonicName, bca = FALSE, percentile = TRUE, R = 1000)
```

##Neg Flu Bed to Dispo Analysis

Shapiro Test - failed
```{r}
with(flu.neg2, shapiro.test(BedToDispo.mins[MnemonicName == "Influenza (POCT)"]))
with(flu.neg2, shapiro.test(BedToDispo.mins[MnemonicName == "Respiratory PCR Panel"]))
```

Wilcoxon test - significant
```{r}
do.call("rbind", describeBy(flu.neg2$BedToDispo.mins, flu.neg2$MnemonicName)) %>% select(n, mean, median, min, max)
wilcox.test(BedToDispo.mins ~ MnemonicName, data = flu.neg2, alternative = "two.sided", var.equal = FALSE, conf.int = TRUE, conf.level = 0.95)
```

Bootstrapped Median CI
```{r}
library(rcompanion)
flu.neg2.btd <- flu.neg2 %>% filter(!is.na(BedToDispo.mins))
groupwiseMedian(data=flu.neg2.btd, BedToDispo.mins ~ MnemonicName, bca = FALSE, percentile = TRUE, R = 1000)
```
##CXR Analysis
Chi-square - All
```{r}
table(flu2$MnemonicName, flu2$XR_Completed)
chisq.test(flu2$MnemonicName, flu2$XR_Completed, correct = FALSE)
```

Chi-square - Pos
```{r}
table(flu.pos2$MnemonicName, flu.pos2$XR_Completed)
chisq.test(flu.pos2$MnemonicName, flu.pos2$XR_Completed, correct = FALSE)
```

Chi-square - All
```{r}
table(flu.neg2$MnemonicName, flu.neg2$XR_Completed)
chisq.test(flu.neg2$MnemonicName, flu.neg2$XR_Completed, correct = FALSE)
```
