==> 19 - Full Datasets Analysis Pt 2.Rmd <==
---
title: "19 - Full Datasets Analysis Pt 2 - "
output: html_notebook
---

#Setup
```{r}
flu <- readRDS(file = "dfs/alere.flu.pcr.Rda")
library(tidyverse)
library(lubridate)
library(psych)
```


##Modified ILI Flu Setup
Subset by modified ILI
```{r}
#fever, cough, sore throat, flu, cold, body*ache, chills
flu.ili <- flu %>% subset(
  grepl("fever",flu$RFV, ignore.case=TRUE)| 
  grepl("cough",flu$RFV, ignore.case=TRUE)|
  grepl("sore.*throat",flu$RFV, ignore.case=TRUE)|
  grepl("throat.*pain",flu$RFV, ignore.case=TRUE)|
  grepl("flu",flu$RFV, ignore.case=TRUE)|
  grepl("cold",flu$RFV, ignore.case=TRUE)|
  grepl("body.*ache",flu$RFV, ignore.case=TRUE)|
  grepl("chills",flu$RFV, ignore.case=TRUE)
)

do.call("rbind", describeBy(flu.ili$AdmitToDischarge.mins, flu.ili$MnemonicName))
```


#Split into positive and negatives
```{r}
table(flu.ili$MnemonicName, flu.ili$FluOverall)

flu.ili.pos <- flu.ili %>% filter(
  FluOverall == "positive")

flu.ili.neg <- flu.ili %>% filter(
  FluOverall == "negative")

```

#All Flu
##Outlier Removal
```{r}
#Remove predefined outliers (removes values LOS <20 min); 
flu.ili2 <- flu.ili %>% filter(!(AdmitToDischarge.mins < 20))
```


##All flu.ili Sex Comparison
```{r}
flu.ili2$Sex %>% ftable(flu.ili2$MnemonicName)
chisq.test(flu.ili2$MnemonicName, flu.ili2$Sex, correct=FALSE)
```

##All flu.ili Age Comparison

```{r}
do.call("rbind", describeBy(flu.ili2$Age, flu.ili2$MnemonicName))
wilcox.test(Age ~ MnemonicName, data = flu.ili2, alternative = "two.sided", var.equal = FALSE, conf.int = TRUE, conf.level = 0.95)
```

##All flu.ili LOS Analysis
Shapiro Test - failed
```{r}
with(flu.ili2, shapiro.test(AdmitToDischarge.mins[MnemonicName == "Influenza (POCT)"]))
with(flu.ili2, shapiro.test(AdmitToDischarge.mins[MnemonicName == "Respiratory PCR Panel"]))
```

Wilcoxon test - significant
```{r}
do.call("rbind", describeBy(flu.ili2$AdmitToDischarge.mins, flu.ili2$MnemonicName)) %>% select(n, mean, median, min, max)
wilcox.test(AdmitToDischarge.mins ~ MnemonicName, data = flu.ili2, alternative = "two.sided", var.equal = FALSE, conf.int = TRUE, conf.level = 0.95)
```

Bootstrapped Median CI
```{r}
library(rcompanion)
groupwiseMedian(data=flu.ili2, AdmitToDischarge.mins ~ MnemonicName, bca = FALSE, percentile = TRUE, R = 1000)
```

##All flu.ili Bed to Dispo Analysis

Shapiro Test - failed
```{r}
with(flu.ili2, shapiro.test(BedToDispo.mins[MnemonicName == "Influenza (POCT)"]))
with(flu.ili2, shapiro.test(BedToDispo.mins[MnemonicName == "Respiratory PCR Panel"]))
```

Wilcoxon test - significant
```{r}
do.call("rbind", describeBy(flu.ili2$BedToDispo.mins, flu.ili2$MnemonicName)) %>% select(n, mean, median, min, max)
wilcox.test(BedToDispo.mins ~ MnemonicName, data = flu.ili2, alternative = "two.sided", var.equal = FALSE, conf.int = TRUE, conf.level = 0.95)
```

Bootstrapped Median CI
```{r}
library(rcompanion)
flu.ili2.btd <- flu.ili2 %>% filter(!is.na(BedToDispo.mins))
groupwiseMedian(data=flu.ili2.btd, BedToDispo.mins ~ MnemonicName, bca = FALSE, percentile = TRUE, R = 1000)
```



#Pos flu.ili
##Outlier Removal
```{r}
#Remove predefined outliers (removes values LOS <20 min); 
flu.ili.pos2 <- flu.ili.pos %>% filter(!(AdmitToDischarge.mins < 20))
```


##Pos Flu Sex Comparison
```{r}
flu.ili.pos2$Sex %>% ftable(flu.ili.pos2$MnemonicName)
chisq.test(flu.ili.pos2$MnemonicName, flu.ili.pos2$Sex, correct=FALSE)
```

##Pos Flu Age Comparison

```{r}
do.call("rbind", describeBy(flu.ili.pos2$Age, flu.ili.pos2$MnemonicName))
wilcox.test(Age ~ MnemonicName, data = flu.ili.pos2, alternative = "two.sided", var.equal = FALSE, conf.int = TRUE, conf.level = 0.95)
```

##Pos Flu LOS Analysis
Shapiro Test - failed
```{r}
with(flu.ili.pos2, shapiro.test(AdmitToDischarge.mins[MnemonicName == "Influenza (POCT)"]))
with(flu.ili.pos2, shapiro.test(AdmitToDischarge.mins[MnemonicName == "Respiratory PCR Panel"]))
```

Wilcoxon test - 
```{r}
do.call("rbind", describeBy(flu.ili.pos2$AdmitToDischarge.mins, flu.ili.pos2$MnemonicName)) %>% select(n, mean, median, min, max)
wilcox.test(AdmitToDischarge.mins ~ MnemonicName, data = flu.ili.pos2, alternative = "two.sided", var.equal = FALSE, conf.int = TRUE, conf.level = 0.95)
```

Bootstrapped Median CI
```{r}
library(rcompanion)
groupwiseMedian(data=flu.ili.pos2, AdmitToDischarge.mins ~ MnemonicName, bca = FALSE, percentile = TRUE, R = 1000)
```

##Pos Flu Bed to Dispo Analysis

Shapiro Test - failed
```{r}
with(flu.ili.pos2, shapiro.test(BedToDispo.mins[MnemonicName == "Influenza (POCT)"]))
with(flu.ili.pos2, shapiro.test(BedToDispo.mins[MnemonicName == "Respiratory PCR Panel"]))
```

Wilcoxon test - significant
```{r}
do.call("rbind", describeBy(flu.ili.pos2$BedToDispo.mins, flu.ili.pos2$MnemonicName)) %>% select(n, mean, median, min, max)
wilcox.test(BedToDispo.mins ~ MnemonicName, data = flu.ili.pos2, alternative = "two.sided", var.equal = FALSE, conf.int = TRUE, conf.level = 0.95)
```

Bootstrapped Median CI
```{r}
library(rcompanion)
flu.ili.pos2.btd <- flu.ili.pos2 %>% filter(!is.na(BedToDispo.mins))
groupwiseMedian(data=flu.ili.pos2.btd, BedToDispo.mins ~ MnemonicName, bca = FALSE, percentile = TRUE, R = 1000)
```




#Neg Flu
##Outlier Removal
```{r}
#Remove predefined outliers (removes values LOS <20 min); 
flu.ili.neg2 <- flu.ili.neg %>% filter(!(AdmitToDischarge.mins < 20))
```


##Neg Flu Sex Comparison
```{r}
flu.ili.neg2$Sex %>% ftable(flu.ili.neg2$MnemonicName)
chisq.test(flu.ili.neg2$MnemonicName, flu.ili.neg2$Sex, correct=FALSE)
```

##Neg Flu Age Comparison

```{r}
do.call("rbind", describeBy(flu.ili.neg2$Age, flu.ili.neg2$MnemonicName))
wilcox.test(Age ~ MnemonicName, data = flu.ili.neg2, alternative = "two.sided", var.equal = FALSE, conf.int = TRUE, conf.level = 0.95)
```

##Neg Flu LOS Analysis
Shapiro Test - failed
```{r}
with(flu.ili.neg2, shapiro.test(AdmitToDischarge.mins[MnemonicName == "Influenza (POCT)"]))
with(flu.ili.neg2, shapiro.test(AdmitToDischarge.mins[MnemonicName == "Respiratory PCR Panel"]))
```

Wilcoxon test - 
```{r}
do.call("rbind", describeBy(flu.ili.neg2$AdmitToDischarge.mins, flu.ili.neg2$MnemonicName)) %>% select(n, mean, median, min, max)
wilcox.test(AdmitToDischarge.mins ~ MnemonicName, data = flu.ili.neg2, alternative = "two.sided", var.equal = FALSE, conf.int = TRUE, conf.level = 0.95)
```

Bootstrapped Median CI
```{r}
library(rcompanion)
groupwiseMedian(data=flu.ili.neg2, AdmitToDischarge.mins ~ MnemonicName, bca = FALSE, percentile = TRUE, R = 1000)
```

##Neg Flu Bed to Dispo Analysis

Shapiro Test - failed
```{r}
with(flu.ili.neg2, shapiro.test(BedToDispo.mins[MnemonicName == "Influenza (POCT)"]))
with(flu.ili.neg2, shapiro.test(BedToDispo.mins[MnemonicName == "Respiratory PCR Panel"]))
```

Wilcoxon test - significant
```{r}
do.call("rbind", describeBy(flu.ili.neg2$BedToDispo.mins, flu.ili.neg2$MnemonicName)) %>% select(n, mean, median, min, max)
wilcox.test(BedToDispo.mins ~ MnemonicName, data = flu.ili.neg2, alternative = "two.sided", var.equal = FALSE, conf.int = TRUE, conf.level = 0.95)
```

Bootstrapped Median CI
```{r}
library(rcompanion)
flu.ili.neg2.btd <- flu.ili.neg2 %>% filter(!is.na(BedToDispo.mins))
groupwiseMedian(data=flu.ili.neg2.btd, BedToDispo.mins ~ MnemonicName, bca = FALSE, percentile = TRUE, R = 1000)
```

##CXR Analysis
Chi-square - All
```{r}
table(flu.ili2$MnemonicName, flu.ili2$XR_Completed)
chisq.test(flu.ili2$MnemonicName, flu.ili2$XR_Completed, correct = FALSE)
```

Chi-square - Pos
```{r}
table(flu.ili.pos2$MnemonicName, flu.ili.pos2$XR_Completed)
chisq.test(flu.ili.pos2$MnemonicName, flu.ili.pos2$XR_Completed, correct = FALSE)
```

Chi-square - All
```{r}
table(flu.ili.neg2$MnemonicName, flu.ili.neg2$XR_Completed)
chisq.test(flu.ili.neg2$MnemonicName, flu.ili.neg2$XR_Completed, correct = FALSE)
```
